code/requirements.txt
python-dotenv>=1.0.0
azure-data-tables>=12.4.0
azure-storage-blob>=12.0.0
azure-ai-vision-face>=1.0.0b2
Pillow>=9.0.0
piexif>=1.1.0
requests>=2.25.0
